refactor(storage): extract Repository protocol with SQLite fallback
The bot's Database class was hardcoded to asyncpg/Supabase. The
side-by-side test harness needs per-instance SQLite databases but the
bot ignored DB_PATH after the Supabase migration.

- storage/repository.py: Protocol with 6 methods the bot actually uses
- storage/supabase.py: Renamed from Database, implements Repository
- storage/sqlite.py: aiosqlite implementation for testing/local use
- storage/__init__.py: Factory — uses Supabase if SUPABASE_DB_URL is
  set, otherwise falls back to SQLite with DB_PATH
- bot.py: Uses create_repository() instead of Database()
- risk_controller.py: Type hint uses Repository protocol
