+++ live_trading_bot/storage/repository.py
@@ -1,0 +1,18 @@
+"""Storage repository protocol.
+
+Defines the interface that the bot depends on for persistence.
+Implementations: SupabaseRepository (production), SqliteRepository (testing/harness).
+"""
+
+from typing import Optional, Protocol
+
+from .models import Trade, SignalRecord, RiskEvent
+
+
+class Repository(Protocol):
+    async def connect(self) -> None: ...
+    async def close(self) -> None: ...
+    async def insert_trade(self, trade: Trade) -> int: ...
+    async def insert_signal(self, signal: SignalRecord) -> int: ...
+    async def insert_risk_event(self, event: RiskEvent) -> int: ...
+    async def get_daily_pnl(self, symbol: Optional[str] = None) -> float: ...
